requirements_local.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
numpy==1.23.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
